requirements.txt

fastapi
uvicorn
torch
torchvision
Pillow
numpy
scikit-image
python-multipart